Renaming dto and error folders
